+++ src/components/common/ThemeProvider/index.tsx
@@ -1,6 +1,40 @@
 import React, { ReactNode } from 'react';
-import { ConfigProvider } from 'antd';
-import { THEME_CONFIG } from '../../../utils';
+import { ConfigProvider, ThemeConfig } from 'antd';
+import { COLOR } from '../../../utils';
+
+export const THEME_CONFIG: ThemeConfig = {
+  token: {
+    colorPrimary: COLOR.PRIMARY,
+    colorBgBase: COLOR.WHITE,
+    colorLink: COLOR.PRIMARY,
+    colorTextPlaceholder: COLOR.GREY_2,
+    fontSize: 18,
+    borderRadius: 5,
+    controlHeight: 42,
+  },
+  components: {
+    Anchor: {
+      linkPaddingBlock: 8,
+    },
+    Layout: {
+      headerBg: COLOR.WHITE,
+      bodyBg: COLOR.WHITE,
+    },
+    Tabs: {
+      motionDurationMid: '0.1s',
+      motionDurationSlow: '0.1s',
+    },
+    Table: {
+      padding: 12,
+      fontWeightStrong: 500,
+    },
+    Alert: {
+      colorInfoBg: COLOR.ANT_PRIMARY_BG,
+      colorInfoBorder: COLOR.ANT_PRIMARY_BORDER,
+      colorInfo: COLOR.PRIMARY /** info icon */,
+    },
+  },
+};
 
 export const CustomThemeProvider = ({ children }: { children: ReactNode }) => {
   return <ConfigProvider theme={THEME_CONFIG}>{children}</ConfigProvider>;