+++ src/components/product/Disclaimer/Disclaimer.tsx
@@ -2,6 +2,8 @@ import React, { ReactNode, MouseEvent } from 'react';
 import { Alert, Button } from 'antd';
 import styled from 'styled-components';
 
+import { CustomThemeProvider } from '../../common/ThemeProvider';
+
 const DisclaimerAlert = styled(Alert)`
   .ant-btn-link {
     padding: 0;
@@ -23,33 +25,35 @@ export const Disclaimer = ({
   // if the href is internal, we don't want to open it in a new tab
   const isInternalLink = href ? href.startsWith('/') : false;
 
+  const defaultMessage = (
+    <div>
+      By accessing this app you agree to the&nbsp;
+      <Button
+        type="link"
+        // style={btnLinkStyle}
+        href={href}
+        target={isInternalLink ? undefined : '_blank'}
+        onClick={(e) => {
+          if (onDisclaimerClick) {
+            // prevent default behavior of the link and call the callback
+            e.preventDefault();
+            onDisclaimerClick(e);
+          }
+        }}
+      >
+        disclaimer
+      </Button>
+    </div>
+  );
+
   return (
-    <DisclaimerAlert
-      type="warning"
-      showIcon
-      closable
-      message={
-        message || (
-          <div>
-            By accessing this app you agree to the&nbsp;
-            <Button
-              type="link"
-              // style={btnLinkStyle}
-              href={href}
-              target={isInternalLink ? undefined : '_blank'}
-              onClick={(e) => {
-                if (onDisclaimerClick) {
-                  // prevent default behavior of the link and call the callback
-                  e.preventDefault();
-                  onDisclaimerClick(e);
-                }
-              }}
-            >
-              disclaimer
-            </Button>
-          </div>
-        )
-      }
-    />
+    <CustomThemeProvider>
+      <DisclaimerAlert
+        type="warning"
+        showIcon
+        closable
+        message={message || defaultMessage}
+      />
+    </CustomThemeProvider>
   );
 };