+++ HISTORY.md
@@ -126,3 +126,9 @@
 
 ## 0.4.10
 - add more functionality to `<AddressLink />` component
+
+## 0.4.11
+- add more functionality for improving `Tokenomics` App
+- update `antd` theme especially for `Alert` component
+- `useScreen` hook
+- date format function such as `getFormattedDate`, `getFullFormattedDate`, `getCommaSeparatedNumber`